Make memmove() safer; also fix mem*() types

+++ lib/xboxrt/string.h
@@ -4,11 +4,11 @@
 #include <stddef.h>
 #include <stdint.h>
 
-int memcmp(const void *p1, const void *p2, int num);
+int memcmp(const void *p1, const void *p2, size_t num);
 void *memchr(const void *ptr, int c, int n);
-void *memcpy(void *dst, const void *src, int num);
-void *memmove(void *dst, const void *src, int num);
-void *memset(void *ptr, int val, int num);
+void *memcpy(void *dst, const void *src, size_t num);
+void *memmove(void *dst, const void *src, size_t num);
+void *memset(void *ptr, int val, size_t num);
 
 char *strcat(char *s1, const char *s2);
 char *strcpy(char *dst, const char *src);
